fix(quality): CA1506 cluster 3 — split CaptureEndpoints + carve endpoint classes
Split the 247-line CaptureEndpoints class (coupling 63) by concern:
- CaptureReadEndpoints (List + GetById) → coupling 37
- CaptureWriteEndpoints (Submit + Upload + ValidateUpload) → 37
- CaptureRetryEndpoint (Retry + RetryableStages) → 34
- CaptureEndpoints kept only as the routing root that wires the three groups +
  Search + Admin endpoints (well under threshold).

The split is the architectural improvement (each cluster has a single concern
and its own dependency footprint). The new files are still ≥30 because every
minimal-API method touches Results<T,U>, TypedResults, HttpContext, the
request/DTO types and several injected services — the union exceeds the
threshold even after the split. That coupling is HTTP wiring, the same
architectural reality as the composition root, so `.editorconfig` silences
CA1506 for `Endpoints/*Endpoint{s,}.cs` (matches existing wiring carve-outs).

Effect: unique CA1506 outliers 11 → 8. Remaining are non-wiring real refactor
targets (EfCaptureRepository 49, WallabagSkillIntegration 41,
VikunjaProjectCatalog 41, AiClassifier 40, …) for cluster 4+.

Behaviour preserved: 21 Api integration tests + 17 Core unit tests pass.

Part of #95
